Extensive changes made to the preferences system.
The menu for editing the preferences is now entirely derived from the
preferences file.  Also the names have been changed to make them
more descriptive and to include units where appropriate.

The system is now also configured for multiple extruders, though the
code to select between them is not yet in.