Started work on the memory management.
destroy() function added to all the geometry classes 
that recursively walks all the other bits of geometry pointed
to by an instance of that class destroy()ing them too.

This is to be used at the end of a layer to get rid of all
the geometry for that layer.

All this is on the assumption that the Java garbage collector 
can't spot a disjoint graph that has cycles of pointers in
it.

Also tuned up the nose-wiping code, and added a call to System.gc()
at the moment where it turns the cooling fan on.