Solenoid valve code added to PIC firmware.  THE FAN NOW WORKS OFF RB6, not the spare channel of the L298.  That channel is used to drive the solenoid.
To drive the fan, wire in a TIP110 to ALT 2, connect in the end of the 220 ohm resistor that goes to the base of that transistor, but connect the other end to hole 4 on the 7-hole connector by the Max/Empty connector.  This goes to RB6.  It's probably a good idea to wire a diode across connector P7 so that it's reverse biased (stripe to the +12v line, in other words); that'll take care of any back EMF from the fan motor. 

The code is compatible with the Arduino, and the Java host software is set up to drive it too.