Code modified to allow the extruder motor to be turned on and off for each line segment, or turned on at the start of a polygon, and off at the end.
The preferences file variable is:

Extruder0_PauseBetweenSegments=true

true makes it pause (the current behaviour, and the default); false
allows it to be on continuously.

I did this as I want to add command buffering to the Arduino, and we probably
dont want the motor going in fits and starts if it is working from 
buffered commands.