Bit of tidying of the way the code handles being run with no RepRap machine connected. It now warns, but doesn't otherwise throw a wobbly.
I have also extended the time that the comms software will wait for an acknowledgement; this
seems to make the Arduino a lot happier.  It can be slow to talk when busy with a buffered 
poly-line.

I have also started modifying the polling thread that updates the temperature etc to
make it easier to control.