Working the control panel back so it will run under Java 1.5, as 1.6 is not available on the Mac.  Some tidying needs to be done, but the majority is now going.
This needs an extra .jar file: swing-layout-1.0.3.jar to be in the class path.  For convenience this is now included in 
the lib directory of the distribution; copy it to (for example) /usr/lib/jvm/java-1.5.0-sun/jre/lib/ext/ so that the 
system can find it.  The need for this file will go away at Java 1.6, as its equivalent is included in the standard system.