This change introduces a potential version 2 stepper driver based on the Allegro A3977 chip. This designed is based on the Stepper controller that was being designed for the Fabr reprap derivative and represents a unification of that effort with RepRap.
This checkin has a mostly complete schematic for the driver itself, using RepRap v2 standards including Power connector, IDC with standard driver pinout, and the .156" headers for the stepper driver.

The schematic still needs to include the end stop connectors, Opto-isolation and isolation diodes. Since we decided to expose the SR as a toggle, I need to determine how to add the external Schottky diodes.

Once reviews are complete on the schematic, I'll work on the board layout.