GCode interpreter set up so it can be used on both the Arduino and Sanguino (the default).
On the Arduino the stand-alone test functions and circular interpolation are not supported, as there isn't a lot of room.

Copy the files pins.h.dist and parameters.h.dist to the equivalent .h files and edit them to match your RepRap machine.

To switch, comment out the #define SANGUINO at the head of pins.h.