Code added to support a four-dimensional RepRap; that is a RepRap that runs a 4D dda to drive X, Y, Z and the extruder (which has to be a stepper-motor extruder).
Set the variable FourD in the file reprap.properties to true to switch this on.  The software will then generate G-codes that look like this:

G1 X32.7 Y47.7 Z0.2526 E25.3 F2700.0 ;print segment

which means move to the (x, y, z) point (32.7, 47.7 0.2526) and simultaneously extrude material to a length of 25.3 mm.  Note that the extrude length is cumulative; that is it gets bigger and bigger with each move, until it is re-zeroed with a G92 command.

You can still set the extrude rate independently of the movement in space.  The resulting E values will correspond to whatever extrude rate in mm/minute you set.

The firmware in the directory FourD_GCode_Interpreter will interpret the resulting G-codes correctly.  It will still also run the conventional G-codes as well.

Set FourD to false to generate conventional G codes.