Improving RFO file handling.
There seems to be a Java bug: if you create a temporary directory then fill it with files and flag both them and the directory using deleteOnExit(), then - on exit - it tries to delete the directory before the files rather than after, and so fails.  How dumb is that?  I will investigate further...