CarapaceCopier: took out some unused methods from Main (point cloud -> DeWall tetrahedrons-> triangles). Added in functionality to take an axis-aligned bounding box and backproject it into an image and set the pixels either inside or outside to 'processed'. This uses projections of the corner points and the 12 triangles the 6 faces are composed of to do quick barycentric coordinate transform and see if the point is inside or outside the triangle. Also changed the progress bar for the triangluation to show progress through the surface voxels (not through the z layers of voxels as before).
Note that the process is a lot slower than it used to be. This has been the case since the revisions 3818 and 3820 that changed the way the images are stored. Presumably as there is more memory being used the garabage collector running in the background is  working harder and any processing dealing with the image class drags around all the additional data as well.