CarapaceCopier: Changed the PixelColour class to have an export to int method and a constructor to accept a previously exported int for reconstruction.
This meant that I could change the internal storage of the pixel colours to a simple 2d int array with the colours being reconstructed to PixelColour when they need to be processed. This saves a significant amount of time compared to storing them as a 2d PixelColour array and does away with the kludge introduced in the previous update.