Added capability for using two different temperature sensors on a Mega style Machine with both an extruder and heated bed.
HINT: To decide which Thermistor-s or Temp Sensors are used, you no longer need to change Temperature.h!
please just change the two definitions in your configuration.h file that are:
TEMP_SENSOR       ( choose what to set it to based on the options in features.h )
BED_TEMP_SENSOR    ( as per above, or leave it unset, and it will automatically be set the same as TEMP_SENSOR )

I also did a cleanup of all the thermistor definitions in the file, documenting the actual part numbers, and supplier info where they were known, and disabling those that I could not identify.